refactor: extract runLaunchFlow into flows/launch-flow.ts
Step 2 of specs/SPEC_AGENT_VIEW_MODULARIZATION_2026_04_13.md.

Scope:
  - New file: frontend/app/view/agent/flows/launch-flow.ts (240 lines)
    * LogFn type export
    * LaunchFlowOptions, LaunchFlowResult types
    * runLaunchFlow(opts) — refactored from positional args to a
      single options object for readability + easier future
      extension (e.g. adding authEnv variants, extra hooks)
  - agent-view.tsx:
    * Remove inline runLaunchFlow function (191 lines)
    * Add `import { runLaunchFlow } from "./flows/launch-flow"`
    * Update the one call site to pass an options object instead
      of 7 positional args

Line counts:
  - agent-view.tsx: 1,053 → 854 (−199)
  - flows/launch-flow.ts: new, 240 lines (+41 vs extracted, from
    imports + options-object type definitions)

Progress toward target:
  Step 0 (baseline): 1,178 lines
  Step 1 done:       1,053 (−125, AgentPicker extracted)
  Step 2 done:       854   (−199, launch flow extracted)
  Remaining:         554   (−554 more needed to hit ≤300)

Behavior change: zero. The launch flow function is byte-identical
inside its new file — only the parameter binding changed from
positional to object-destructured, which is a compile-time no-op.
All RpcApi calls, all waveEventSubscribe subscriptions, all state
transitions, all log sink invocations unchanged.

Call-site refactor: the old positional call was
    runLaunchFlow(model.blockId, prov, log, setAuthUrl,
                  () => loginCancelled, setLoginWaiting, authEnv)
The new call is
    runLaunchFlow({ blockId, provider, log, setAuthUrl,
                    isCancelled, setLoginWaiting, authEnv })
— same arguments, same order of side effects, better readability
for future hook extractions that will pass pieces of context through.
